+++ src/_FastBSplineBasis.jl
@@ -1,11 +1,12 @@
-function bsplinebasis(i::Int,P::FastBSplineSpace{0},t::Real)
+# TODO: use macro for bsplienbasis(::FastBSplineSpace)
+function bsplinebasis(i::Integer,P::FastBSplineSpace{0},t::Real)
     k_1, k_2, k_end = P.vector[i], P.vector[i+1], P.vector[end]
     B_1 = Float64((k_1 ≤ t < k_2) | (k_1 ≤ t ≤ k_2 == k_end))
 
     return B_1
 end
 
-function bsplinebasis(i::Int,P::FastBSplineSpace{1},t::Real)
+function bsplinebasis(i::Integer,P::FastBSplineSpace{1},t::Real)
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
     k_1, k_2, k_3, k_end = P.vector[i], P.vector[i+1], P.vector[i+2], P.vector[end]
     B_1, B_2 = Float64(k_1 ≤ t < k_2), Float64((k_2 ≤ t < k_3) | (k_2 ≤ t ≤ k_3 == k_end))
@@ -16,7 +17,7 @@ function bsplinebasis(i::Int,P::FastBSplineSpace{1},t::Real)
     return B_1
 end
 
-function bsplinebasis(i::Int,P::FastBSplineSpace{2},t::Real)
+function bsplinebasis(i::Integer,P::FastBSplineSpace{2},t::Real)
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
     k_1, k_2, k_3, k_4, k_end = P.vector[i], P.vector[i+1], P.vector[i+2], P.vector[i+3], P.vector[end]
     B_1, B_2, B_3 = Float64(k_1 ≤ t < k_2), Float64(k_2 ≤ t < k_3), Float64((k_3 ≤ t < k_4) | (k_3 ≤ t ≤ k_4 == k_end))
@@ -30,7 +31,7 @@ function bsplinebasis(i::Int,P::FastBSplineSpace{2},t::Real)
     return B_1
 end
 
-function bsplinebasis(i::Int,P::FastBSplineSpace{3},t::Real)
+function bsplinebasis(i::Integer,P::FastBSplineSpace{3},t::Real)
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
     k_1, k_2, k_3, k_4, k_5, k_end = P.vector[i], P.vector[i+1], P.vector[i+2], P.vector[i+3], P.vector[i+4], P.vector[end]
     B_1, B_2, B_3, B_4 = Float64(k_1 ≤ t < k_2), Float64(k_2 ≤ t < k_3), Float64(k_3 ≤ t < k_4), Float64((k_4 ≤ t < k_5) | (k_4 ≤ t ≤ k_5 == k_end))
@@ -47,11 +48,11 @@ function bsplinebasis(i::Int,P::FastBSplineSpace{3},t::Real)
     return B_1
 end
 
-function bsplinebasis′₊₀(i::Int,P::FastBSplineSpace{0},t::Real)
+function bsplinebasis′₊₀(i::Integer,P::FastBSplineSpace{0},t::Real)
     return 0.0
 end
 
-function bsplinebasis′₊₀(i::Int,P::FastBSplineSpace{1},t::Real)
+function bsplinebasis′₊₀(i::Integer,P::FastBSplineSpace{1},t::Real)
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
     k_1, k_2, k_3, k_end = P.vector[i], P.vector[i+1], P.vector[i+2], P.vector[end]
     B_1, B_2 = Float64(k_1 ≤ t < k_2), Float64(k_2 ≤ t < k_3)
@@ -62,7 +63,7 @@ function bsplinebasis′₊₀(i::Int,P::FastBSplineSpace{1},t::Real)
     return B_1
 end
 
-function bsplinebasis′₊₀(i::Int,P::FastBSplineSpace{2},t::Real)
+function bsplinebasis′₊₀(i::Integer,P::FastBSplineSpace{2},t::Real)
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
     k_1, k_2, k_3, k_4, k_end = P.vector[i], P.vector[i+1], P.vector[i+2], P.vector[i+3], P.vector[end]
     B_1, B_2, B_3 = Float64(k_1 ≤ t < k_2), Float64(k_2 ≤ t < k_3), Float64(k_3 ≤ t < k_4)
@@ -76,7 +77,7 @@ function bsplinebasis′₊₀(i::Int,P::FastBSplineSpace{2},t::Real)
     return B_1
 end
 
-function bsplinebasis′₊₀(i::Int,P::FastBSplineSpace{3},t::Real)
+function bsplinebasis′₊₀(i::Integer,P::FastBSplineSpace{3},t::Real)
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
     k_1, k_2, k_3, k_4, k_5, k_end = P.vector[i], P.vector[i+1], P.vector[i+2], P.vector[i+3], P.vector[i+4], P.vector[end]
     B_1, B_2, B_3, B_4 = Float64(k_1 ≤ t < k_2), Float64(k_2 ≤ t < k_3), Float64(k_3 ≤ t < k_4), Float64(k_4 ≤ t < k_5)
@@ -97,7 +98,7 @@ end
 """
 TODO: faster.....
 """
-function bsplinebasis′(i::Int64, P::FastBSplineSpace, t::Real)::Float64
+function bsplinebasis′(i::Integer, P::FastBSplineSpace, t::Real)::Float64
     ÷(a,b) = ifelse(b == 0.0, 0.0, a/b)
 
     p = degree(P)
